:gift: feat: 반응형 ui 구현
