+++ src/app/layout.tsx
@@ -4,7 +4,6 @@ import "@/app/globals.css";
 import { LikeContextProvider } from "@/like-context";
 
 const inter = Inter({ subsets: ["latin"] });
-// TODO: app의 layout - 여기 보기
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="kr">
