+++ src/app/ui/common/header.module.css
@@ -1,13 +1,15 @@
 .header {
   width: 100%;
-  height: 50px;
+  min-height: 50px;
   display: flex;
   justify-content: space-between;
+  flex-wrap: wrap;
   align-items: center;
   padding: 0 50px;
 }
 
 .buttonWrapper {
+  flex-wrap: wrap;
   display: flex;
   gap: 12px;
 }
@@ -34,6 +36,7 @@
 
 .userWrapper {
   display: flex;
+  flex-wrap: wrap;
   align-items: center;
   color: gray;
   font-weight: 600;
