+++ src/app/ui/common/modal.module.css
@@ -1,8 +1,3 @@
-.modalRef {
-  width: 500px;
-  height: 500px;
-}
-
 .backdrop {
   min-width: 1510px;
   z-index: 1500;
@@ -27,7 +22,7 @@
   flex-direction: column;
   background-color: #fff;
   width: 80%;
-  height: 80%;
+  /* height: 80%; */
 }
 
 .modalTitle {
@@ -81,8 +76,7 @@
 
 .modalImage {
   width: 100%;
-  height: 100%;
-  max-height: 400px;
+  height: 45vh;
   margin: 30px auto 0;
   position: relative;
 }
