+++ src/app/ui/common/pagination.tsx
@@ -1,7 +1,7 @@
-import { useSearchParams, usePathname } from "next/navigation";
-import Link from "next/link";
-import { ChevronRightIcon, ChevronLeftIcon, EllipsisHorizontalIcon } from "@heroicons/react/24/outline";
 import clsx from "clsx";
+import Link from "next/link";
+import { useSearchParams, usePathname } from "next/navigation";
+import { ChevronRightIcon, ChevronLeftIcon, EllipsisHorizontalIcon } from "@heroicons/react/24/outline";
 import styles from "@/app/ui/common/pagination.module.css";
 
 type Pagination = {
