+++ src/app/ui/common/photo-list.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import PhotoItem from "@/app/ui/common/photo-item";
-import styles from "@/app/ui/common/photo-list.module.css";
 import clsx from "clsx";
+import PhotoItem from "@/app/ui/common/photo-item";
 import Pagination from "@/app/ui/common/pagination";
-import { PAGINATION } from "@/constants/index";
 import { useSearchParams } from "next/navigation";
 import { Photo } from "@/app/lib/definitions";
+import { PAGINATION } from "@/app/constants";
+import styles from "@/app/ui/common/photo-list.module.css";
 
 type PhotoListType = {
   photoData: { results: Photo[]; total_pages: number };
@@ -16,10 +16,7 @@ const PhotoList = ({ photoData }: PhotoListType) => {
   const { results: photos = [], total_pages } = photoData;
   const params = useSearchParams();
   const page = Number(params.get("page")) || 1;
-
-  if (photos.length === 0) {
-    return <p>사진이 없습니다.</p>;
-  }
+  if (photos.length === 0) return <p>사진이 없습니다.</p>;
 
   return (
     <section>
