+++ src/app/ui/search.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { MagnifyingGlassIcon } from "@heroicons/react/24/outline";
-import styles from "@/app/ui/search.module.css";
-import { useDebouncedCallback } from "use-debounce";
 import clsx from "clsx";
+import { MagnifyingGlassIcon } from "@heroicons/react/24/outline";
 import { useSearchParams, usePathname, useRouter } from "next/navigation";
+import { useDebouncedCallback } from "use-debounce";
+import styles from "@/app/ui/search.module.css";
 
 const SearchForm = () => {
   const searchParam = useSearchParams();
